requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.1
python-dotenv==1.0.0

# External APIs
google-cloud-aiplatform==1.38.0
cohere==5.15.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3
rich==13.7.0